feat: database create 로직을 lib/ 모듈로 분리하고 타입관련 오류 수정
