+++ server/package.json
@@ -9,7 +9,7 @@
     "start": "cross-env NODE_ENV=development nodemon --exec ts-node -r tsconfig-paths/register src/server.ts",
     "test": "cross-env NODE_ENV=development jest",
     "lint": "eslint .",
-    "db:create": "cross-env NODE_ENV=development ts-node src/config/config --create"
+    "db:create": "cross-env NODE_ENV=development ts-node src/lib/createDatabase.ts"
   },
   "lint-staged": {
     "src/**/*.{js,ts}": [
